FwCloud entity
